댓글 추천 기능 오류 수정, 일부 엔티티 칼럼 수정 및 삭제

+++ src/main/java/com/yk/web/WebRestController.java
@@ -22,7 +22,7 @@ public class WebRestController {
 	
 	
     @PostMapping("/registration")
-    public int userRegistration(@Valid @RequestBody UserRequestDto UserDto) {
+    public long userRegistration(@Valid @RequestBody UserRequestDto UserDto) {
     	return userServiceImpl.userResistrationPro(UserDto);
     }
