+++ src/main/java/com/yk/web/user/dao/UserRepository.java
@@ -23,5 +23,5 @@ public interface UserRepository extends JpaRepository<Users, Long>{
 
     @Modifying
     @Query("update Users u set u.password = :password where u.id = :id")
-    void updatePassword(@Param("password") String password, @Param("id") int id);
+    void updatePassword(@Param("password") String password, @Param("id") long id);
 }
